Changed project structure to prevent code dupilcates
